fix: tile completion state was not being stored with statePersist.
In order to store completion states into local storage, the states had to be
moved to the Chessmite component (and passed into the Tile component) instead
of having them created there. This was in order to ensure that each state could
be uniquely indexed when stored and retrieved from localStorage.